Remove deprecations from AppContainer
In the alpha versions of 3.0, AppContainer accepted the child component and
its props as props, but we moved to passing a child in the beta. The beta's been out for a while, and I haven't seen any examples
recently using the alpha examples, so I'd say it's safe to remove these deprecations.